requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.1
    # via django
attrs==21.4.0
    # via pytest
beautifulsoup4==4.11.1
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
bleach==5.0.0
    # via django-markdownfield
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage==6.3.2
    # via -r requirements.in
cryptography==37.0.2
    # via pyopenssl
django==4.0.4
    # via
    #   -r requirements.in
    #   django-contrib-comments
    #   django-extensions
    #   django-js-asset
    #   django-markdownfield
    #   django-markup
    #   django-recaptcha
django-autoslug==1.9.8
    # via -r requirements.in
django-contrib-comments==2.2.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-mptt
django-markdownfield==0.10.0
    # via -r requirements.in
django-markup==1.5
    # via -r requirements.in
django-mptt==0.13.4
    # via -r requirements.in
django-recaptcha==3.0.0
    # via -r requirements.in
django-richtextfield==1.6.1
    # via -r requirements.in
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
lxml==4.8.0
    # via -r requirements.in
markdown==3.3.6
    # via django-markdownfield
mypy-extensions==0.4.3
    # via black
mysqlclient==2.1.0
    # via -r requirements.in
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pdfkit==1.0.0
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pillow==9.1.0
    # via -r requirements.in
pip-tools==6.6.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via -r requirements.in
pyopenssl==22.0.0
    # via -r requirements.in
pyparsing==3.0.8
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-dotenv
pytest-django==4.5.2
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via
    #   -r requirements.in
    #   pytest-dotenv
requests==2.27.1
    # via -r requirements.in
sentry-sdk==1.5.11
    # via -r requirements.in
shortuuid==1.0.8
    # via django-markdownfield
six==1.16.0
    # via bleach
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via
    #   black
    #   pep517
    #   pytest
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
webencodings==0.5.1
    # via bleach
werkzeug==2.1.2
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
